Initial implementation of atttack-defense tree
Currently assumes defense nodes have the same format as attack nodes.
Impact is equivalent to cost of the defense node, riskIndex isn't used
at the moment. Currently outputs in the same format as the attack nodes
with each scenario corresponding to a successful combination of defense
nodes.

Game-theoretic equilibrium not implemented yet.